Move AudioFeedback into new file to share code with Mac app
