+++ TheatricalMovieTrailers/Views/CoverFlowScrollView.swift
@@ -306,50 +306,3 @@ struct CoverFlowScrollView_Previews: PreviewProvider {
     }
 }
 #endif
-
-// MARK: - UI Audio Feedback
-fileprivate class AudioFeedback {
-    private let selectAudioPlayer: AVAudioPlayer?
-    private let limitAudioPlayer: AVAudioPlayer?
-    private let exitAudioPlayer: AVAudioPlayer?
-    private let selectionChangeAudioPlayer: AVAudioPlayer?
-    
-    init() {
-        selectAudioPlayer = try? AVAudioPlayer(contentsOf: Bundle.main.url(forResource: "Selection", withExtension: "aif")!)
-        limitAudioPlayer = try? AVAudioPlayer(contentsOf: Bundle.main.url(forResource: "Limit", withExtension: "aif")!)
-        exitAudioPlayer = try? AVAudioPlayer(contentsOf: Bundle.main.url(forResource: "Exit", withExtension: "aif")!)
-        selectionChangeAudioPlayer = try? AVAudioPlayer(contentsOf: Bundle.main.url(forResource: "SelectionChange", withExtension: "aif")!)
-        #if os(iOS)
-        setAudioPriority(playingTrailer: false)
-        #endif
-    }
-    
-    private func tryPlay(_ player:AVAudioPlayer?) {
-        if Settings.instance.playUISounds, let player = player {
-            DispatchQueue.global(qos: .userInitiated).async {
-                player.play()
-            }
-        }
-    }
-    
-    public func selection() {
-        tryPlay(selectAudioPlayer)
-    }
-    
-    public func limit() {
-        tryPlay(limitAudioPlayer)
-    }
-    
-    public func exit() {
-        tryPlay(exitAudioPlayer)
-    }
-    
-    public func selectionChange() {
-        tryPlay(selectionChangeAudioPlayer)
-    }
-    
-    @available(iOS 14, *)
-    public func setAudioPriority(playingTrailer: Bool) {
-        try? AVAudioSession.sharedInstance().setCategory(playingTrailer ? .playback : .ambient)
-    }
-}
